gg.md
In this code, the user is prompted to enter a year using the input() function, and the input is converted to an integer using the int() function.

The year is then checked using an if statement. According to the leap year rules, if a year is divisible by 4 and not divisible by 100, or if it is divisible by 400, then it is a leap year. If the conditions are satisfied, a corresponding message is printed indicating that the year is a leap year. Otherwise, a different message is printed indicating that it is not a leap year.